Do not change the PC when accessing memory
Unicorn does not allow to change the PC during a memory access. When the
simulation is suspended during a memory access (e.g., due to a bp hit),
all registers are flushed when resumed to forward possible changes of
the register property values to the ISS. In this case, an error occurs
when the PC should be updated. To overcome this issue, the PC update is
not forwarded when the value is not changed. When a memory access is
ongoing, the value is not updated.

Signed-off-by: Nils Bosbach <bosbach@ice.rwth-aachen.de>